+++ src/components/Persona/Persona.tsx
@@ -3,25 +3,7 @@ import {Icon} from '../Icon';
 import {Mail} from '../icons';
 import {PersonaButton, PersonaWrap} from '../PersonaWrap';
 import {getTwoLetters} from './getTwoLetters';
-
-export interface PersonaProps {
-    /** Visible text */
-    text: string;
-    /** Image source */
-    image?: string;
-    /** Visual appearance (with or without border) */
-    theme?: 'default' | 'clear';
-    /** Avatar appearance */
-    type?: 'person' | 'email' | 'empty';
-    /** Text size */
-    size?: 's' | 'n';
-    /** Handle click on button with cross */
-    onClose?: (text: string) => void;
-    /** Handle click on component itself */
-    onClick?: (text: string) => void;
-    /** Custom CSS class for root element */
-    className?: string;
-}
+import type {PersonaProps} from './types';
 
 const PersonaComponent: FC<PersonaProps> = ({
     size = 's',
