+++ src/components/Persona/index.ts
@@ -1,1 +1,2 @@
 export * from './Persona';
+export type {PersonaProps} from './types';
